fix: changed position of sort dropdown
